requirements.txt
# Core data
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
pyarrow>=15.0.0

# Geospatial
h3>=4.1.0
shapely>=2.0.0
geopy>=2.4.0
pydeck>=0.9.0

# ML / NLP
scikit-learn>=1.4.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0

# LLM / Agents
openai>=1.30.0
pydantic>=2.6.0

# Observability
mlflow>=2.20.0

# UI
streamlit>=1.32.0
streamlit-folium>=0.19.0
folium>=0.16.0
plotly>=5.20.0
keplergl>=0.3.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.7.0
tenacity>=8.2.3
orjson>=3.10.0
